start of refactor to convert hardcoded values to variables
to pull this off you need the ability to pass vars to screen,
defvar-local, a=${a:-b} and other bash assignment hacks, etc. how many
layers of indirection do you need when passing variables between
runtimes? the answer seems to be n, where n is the number of different
pieces of software between the source and target runtime plus the
source and the target ... sigh