cli build-image now reraises errors from blocks
this completes the flow control passing needed to use source.org to
compose and run arbitrary blocks e.g. sckan-build-save-image is now
maintained as a single process with full flow control by composing
the individual parts into a single larger whole, it is encapsulated
behind the hacked tag run:sckan-build-save-image which means that we
we can maintain and update the process in a single place without
having to fiddle with updating and fixing things in other files e.g.
the sparc developer guide (which inspired this use case)