test emacs kg-dev-user, work on prcl, start on tgbugs-dev-user
tgbugs-dev-user provides a reference development environment that more
or less matches the user environment that most of my work is done,
there are a couple other places where similar work has been done and
they are referenced, but those places are pretty much always a subset
of all the projects so there hasn't been anywhere with complete docs,
probably also missing some stuff needed for e.g. my dissertation,
though iirc I was trying to minimize the env needed for that and
use an ebuild and a local repo to directly specify the deps

some of the kg-dev-user build has been broken up so that it can be
reused in tgbugs-dev-user, there are also issues here with the sbcl ql
https bootstrap process that are not resolved as part of this commit
but will be in the near future

sbcl-stripped fixed up so the work for ql https can be reused in
kg-dev-user and elsewhere, still not do, but it is more or less
working, had to make a fix for the profile changes so that /lib is
correctly symlinked to usr/lib otherwise we see an exec /usr/bin/sbcl:
no such file or directory error

builder-vars don't overwrite _path_dockerfiles if it was already set

test-image add --debug option to simplify matching exact (ish) state
when something goes wrong

in the context of testing if we are mounting dockerfiles the question
is whether we even need to tangle stuff to ./bin/test/ or whether we
just tangle directly to test, the issue devolves to whether the image
name matches the heading matches the folder inside musl etc.